requirements.txt
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
asgiref==3.9.1
asttokens==3.0.0
certifi==2025.7.9
cfgv==3.4.0
decorator==5.2.1
distlib==0.3.9
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
executing==2.2.0
filelock==3.18.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
ipdb==0.13.13
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
Markdown==3.8.2
matplotlib-inline==0.1.7
nodeenv==1.9.1
openai==1.95.1
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
pre_commit==4.2.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
PyYAML==6.0.2
setuptools==78.1.1
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
virtualenv==20.31.2
wcwidth==0.2.13
wheel==0.45.1
